Made MainComponent own file. Moved main react call into app.js
